vp9: decode profile 2 (10-bit 4:2:0) byte-exact vs libvpx 1.15
Store every plane as u16 and thread the frame's bd through the pixel
kernels (clip_pixel_bd, bd-scaled loop-filter thresholds, base =
128 << (bd-8) intra fills, highbd idct guard), so one decode walk serves
8- and 10-bit. The one depth-dependent bitstream element is the CAT6
coefficient token: 14 bits/vp9_cat6_prob at 8-bit, 16 bits/
vp9_cat6_prob_high12 + 2 at 10/12-bit -- ported.

Profile gate now accepts 0 and 2; profiles 1/3 and 12-bit are refused
by name; non-4:2:0 subsampling stays refused.

Byte-exact vs libvpx 1.15 on both 10-bit corpus streams (48 and 120
frames) and the whole 8-bit regression corpus. cargo test -p ec-vp9
green; warning parity (the 5 pre-existing). See lanes/vp9hbd.report.md.

+++ crates/ec-vp9/tests/dump_probe.rs
@@ -17,7 +17,7 @@ fn diff_stats() {
     let mut first: Option<(usize, usize)> = None;
     for y in 0..240 {
         for x in 0..320 {
-            let a = pic.y[y * pic.stride + x];
+            let a = pic.y[y * pic.stride + x] as u8;
             let b = refout[y * w + x];
             let (sy, sx) = (y / 64, x / 64);
             sb_total[sy][sx] += 1;
